requirements.txt

streamlit
python-dotenv
PyPDF2
pdf2image
pytesseract
langchain
faiss-cpu  
transformers  
langchain-google-genai
google-generativeai
